src/requirements.txt
requests
schedule
lxml
PyYAML
jinja2